requirements.txt
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.3.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==2.2.0
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.5
pytz==2025.2
sniffio==1.3.1
SQLAlchemy==2.0.43
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
